Added Run Branch SQL
Added a “module” to check to see if there is an SQL file with the same
name as your branch. If there is, it will check to see if you want to
run it.
If you do, it will check to see if you have php56-ssh2 installed. If
you don’t, it will brew it. If that doesn’t work it will continue.
Otherwise, your SQL will be executed.

If the output from the mysql call comes back with “ERROR” in it, it
will display the error message, otherwise it assumes all is well in the
world.

Also added a quite little countdown method that sleeps and outputs a
message.